fixed goal update and delete auth

+++ backend/controllers/goalController.js
@@ -1,5 +1,6 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // @desc Get goals
 // @route GET /api/goals
@@ -30,9 +31,18 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found");
   }
+  const user = await User.findById(req.user.id)
+
+  // Make sure logged in user matches goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
+
   res.status(200).json(updatedGoal);
 });
 
@@ -45,6 +55,15 @@ const deleteGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found");
   }
+
+  const user = await User.findById(req.user.id)
+
+  // Make sure logged in user matches goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+
   const deletedGoal = await Goal.findByIdAndDelete(req.params.id);
   res.status(200).json(deletedGoal);
 });
